docs : "게시물 관련 기능 수정"
Post, User Class,Controller,Service 수정

+++ MSL-back/src/main/java/com/maswilaeng/service/PostService.java
@@ -6,6 +6,8 @@ import com.maswilaeng.domain.repository.PostRepository;
 import com.maswilaeng.domain.repository.UserRepository;
 import com.maswilaeng.dto.post.request.PostRequestDto;
 import com.maswilaeng.dto.post.request.PostUpdateDto;
+import com.maswilaeng.dto.post.response.UserPostListResponseDto;
+import com.maswilaeng.utils.UserContext;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
@@ -14,6 +16,9 @@ import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import javax.persistence.EntityNotFoundException;
+import javax.xml.bind.ValidationException;
+import java.util.List;
 import java.util.Objects;
 
 @Slf4j
@@ -24,25 +29,19 @@ public class PostService {
     private final PostRepository postRepository;
     private final UserRepository userRepository;
 
-    /* CREATE */
     @Transactional
-    public void save(Long userId, PostRequestDto PostRequestDto) {
-
-        /* User 정보를 가져와 dto에 담아준다. */
+    public void save(Long userId, PostRequestDto postRequestDto) {
         User user = userRepository.findById(userId).get();
-        postRepository.save(PostRequestDto.toEntity(user));
+        postRepository.save(postRequestDto.toEntity(user));
     }
 
-    /* READ 게시글 리스트 조회 */
     @Transactional(readOnly = true)
     public Post findPostById(Long postId) {
         return postRepository.findById(postId).orElseThrow(() ->
-                new IllegalArgumentException("해당 게시글이 존재하지 않습니다. id: " + postId));
+                new EntityNotFoundException("해당 게시글이 존재하지 않습니다. id: " + postId));
     }
 
-    /*
-     * UPDATE -> dirty checking 으로 하는지?
-     */
+
     @Transactional
     public void updatePost(Long userId, PostUpdateDto updateDto) throws Exception { // id 없는 객체 -> null "mergeX"
 
@@ -58,24 +57,20 @@ public class PostService {
 
     /* DELETE */
     @Transactional
-    public void delete(Long id) {
-        Post post = postRepository.findById(id).orElseThrow(() ->
-                new IllegalArgumentException("해당 게시글이 존재하지 않습니다. id=" + id));
+    public void delete(Long userId, Long postId) throws ValidationException {
+        Post post = postRepository.findById(postId).orElseThrow(() ->
+                new IllegalArgumentException("해당 게시글이 존재하지 않습니다. id=" + postId));
 
+        if (!Objects.equals(userId, post.getUser().getId())) {
+            throw new ValidationException("접근 권한 없음");
+        }
         postRepository.delete(post);
     }
 
-    public Page<Post> getUserPostList(Long userId, int currentPage) {
-        return postRepository.findByUserId(userId, PageRequest.of(
-                currentPage - 1, 20, Sort.Direction.DESC, "createdAt"));
+    public UserPostListResponseDto getUserPostList(Long userId) {
+        return new UserPostListResponseDto(postRepository.findAllPostByUserId(userId));
     }
 
-//
-//    public List<PostListResponseDto> getPostListDefault() {
-//        return postRepository.findAllFetchJoin(PageRequest.of(0,500));
-//
-//    }
-//
 //    /* search */
 //    @Transactional(readOnly = true)
 //    public Page<Post> search(String keyword, Pageable pageable) {
